chore(router): removed sql routing store
Removed the sql routing store for the page router plugin. The sql store was not used or supported well, so it was removed due to it's size and complexity. Routes don't often change and therefor database loading is not something that is likely to be part of the core framework.