feat(router): Added ability to ignore xml route format errors when loading routes
This features adds the ability for users to allow format errors in the XML route objects that cause an invalid route to be deserialzied, but load remaining routes with best effort, and clearly logging the errors as warning log event messegs. The property is ignore_errors, and defaults to true. When set to `false` will cause the application to fail loading routes and raise exceptions to the server process causing http 500 errors.